pyproject.toml
[tool.poetry]
name = "tutorials"
version = "0.1.0"
description = "tutorials for integrating python with SQl"
authors = ["liu chiko"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector-python = "^8.0.27"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
